tech_interview_assistant/pyproject.toml
[tool.poetry]
name = "tech-interview-assistant"
version = "0.1.0"
description = "An agentic system to help gathering DS technical interview questions and serve as knowlege base."
authors = ["Shisi Wang <kavender.wang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
python-dotenv = "^1.0.1"
langchain = "^0.2.13"
langgraph = "^0.2.3"
langchain-openai = "^0.1.21"
langchain-google-community = "^1.0.7"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
google-api-python-client = "^2.140.0"
google-search-results = "^2.4.2"
beautifulsoup4 = "^4.12.3"
ipdb = "^0.13.13"
selenium = "^4.23.1"
unstructured = "^0.15.3"
pandas = "^2.2.2"
llama-index = "^0.10.67.post1"
langchain-community = "^0.2.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
